warning pass, fix many -Wall, -Wextra warnings
 - many signed vs unsigned comparisons
 - some mixed signed/unsigned result ternary expressions
 - (void) cast some unused parameters
 - remove some unused parameters
 - add missing initializers to some structure initializations
 - restructure fill object initialization to only brace initialize
   the base structure
 - add /* FALLTHROUGH */ comments to a few switch statements to
   suppress some switch case fall through warnings

+++ datatypes.c
@@ -14,7 +14,7 @@ btm_new(i_img_dim xsize,i_img_dim ysize) {
   struct i_bitmap *btm;
   btm=(struct i_bitmap*)mymalloc(sizeof(struct i_bitmap)); /* checked 4jul05 tonyc */
   bytes = (xsize*ysize+8)/8;
-  if (bytes * 8 / ysize < xsize-1) { /* this is kind of rough */
+  if (bytes * 8 / ysize < (size_t)xsize-1) { /* this is kind of rough */
     fprintf(stderr, "Integer overflow allocating bitmap (" i_DFp ")",
 	    i_DFcp(xsize, ysize));
     exit(3);
